added user authentication to blog

+++ rails/blog/app/controllers/comments_controller.rb
@@ -1,9 +1,12 @@
 class CommentsController < ApplicationController
+  before_action :authenticate_user, only: [:create, :destroy]
+
   def create
     comment_params = params.require(:comment).permit(:body)
     @p = Post.find params[:post_id]
     @comment = Comment.new comment_params
     @comment.post = @p
+    @comment.user = current_user
 
     if @comment.save
       redirect_to post_path(@p), notice: "Comment created succussfully!"
@@ -16,6 +19,7 @@ class CommentsController < ApplicationController
   def destroy
     comment = Comment.find_by_id params[:id]
     if comment
+      redirect_to post_path(params[:post_id]), alert: "Access denied." and return unless can? :destroy, comment
       comment.destroy
       redirect_to post_path(comment.post), notice: "Comment deleted"
     else
